Correction d'une erreur qui entrainer la suppression de l'image même si elle n'etait pas modifiée

+++ controllers/sauce.js
@@ -62,16 +62,22 @@ exports.modifySauce = (req, res, next) => {
             ...JSON.parse(req.body.sauce),
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : {...req.body};
-    Sauce.findOne({ _id: req.params.id })
-        .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
-                    .then(() => res.status(200).json({message: 'Sauce modifiée'}))
-                    .catch(error => res.status(400).json({error}));
-            });
-        })
-        .catch(error => res.status(500).json({ error }));
+    if (req.file) {
+        Sauce.findOne({_id: req.params.id})
+            .then(sauce => {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
+                        .then(() => res.status(200).json({message: 'Sauce modifiée'}))
+                        .catch(error => res.status(400).json({error}));
+                });
+            })
+            .catch(error => res.status(500).json({error}));
+    } else {
+        Sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
+            .then(() => res.status(200).json({message: 'Sauce modifiée'}))
+            .catch(error => res.status(400).json({error}));
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
